+++ packages/attribute-registry-process/package.json
@@ -41,6 +41,7 @@
     "connection-string": "^4.3.6",
     "dotenv-flow": "^3.2.0",
     "express": "^4.18.2",
+    "mongodb": "5.6.0",
     "openapi-zod-client": "^1.7.1",
     "pagopa-interop-commons": "workspace:*",
     "pagopa-interop-models": "workspace:*",
